+++ dio/test/mock/http_mock.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:dio/dio.dart';
 import 'package:mockito/annotations.dart';
 
 import 'http_mock.mocks.dart';
@@ -13,6 +14,7 @@ final httpClientMock = MockHttpClient();
     MockSpec<HttpClientRequest>(),
     MockSpec<HttpClientResponse>(),
     MockSpec<HttpHeaders>(),
+    MockSpec<Transformer>(),
   ],
 )
 class MockHttpOverrides extends HttpOverrides {