+++ dio/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased 6.0.0
 
 - Remove `DefaultHttpClientAdapter` which was deprecated in `5.0.0`.
+- Remove `IOHttpClientAdapter.onHttpClientCreate` which was deprecated in `5.2.0`
 
 ## Unreleased