+++ dio/CHANGELOG.md
@@ -8,7 +8,7 @@
 - `IOHttpClientAdapter.onHttpClientCreate` has been deprecated and is scheduled for removal in
   Dio 6.0.0 - Please use the replacement `IOHttpClientAdapter.createHttpClient` instead.
 - Using `CancelToken` no longer closes and re-creates `HttpClient` for each request when `IOHttpClientAdapter` is used.
-- Fix timeout handling for browsers on `connectionTimeout` and `receiveTimeout`.
+- Fix timeout handling for browsers on `connectTimeout` and `receiveTimeout`.
 
 ## 5.1.2