+++ dio/CHANGELOG.md
@@ -12,6 +12,8 @@ See the [Migration Guide][] for the complete breaking changes list.**
   Dio 6.0.0 - Please use the replacement `IOHttpClientAdapter.createHttpClient` instead.
 - Using `CancelToken` no longer closes and re-creates `HttpClient` for each request when `IOHttpClientAdapter` is used.
 - Fix timeout handling for browser `receiveTimeout`.
+- Using `CancelToken` no longer closes and re-creates `HttpClient` for each request when `IOHttpClientAdapter` is used. 
+- Improve performance when sending binary data (`List<int>`/`Uint8List`). 
 
 ## 5.1.2